refactor and clean up components
