+++ package.json
@@ -7,7 +7,10 @@
     "type": "git",
     "url": "https://github.com/code-for-canada/start-map.git"
   },
-  "keywords": ["street art", "map"],
+  "keywords": [
+    "street art",
+    "map"
+  ],
   "homepage": "https://streetartoronto.ca",
   "dependencies": {
     "@mars/heroku-js-runtime-env": "^3.0.2",
