+++ src/assets/scss/FeatureSlider.scss
