+++ src/assets/scss/Splash.scss
