+++ src/components/App.js
@@ -1,17 +1,7 @@
 import React from 'react';
-// See: https://create-react-app.dev/docs/adding-bootstrap/
-import 'bootstrap/dist/css/bootstrap.min.css'; // Must come first.
-import './App.css';
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import LazyLoad, { forceCheck } from 'react-lazyload';
 import ReactGA from 'react-ga';
-import 'simplebar/dist/simplebar.css';
 import sort from 'fast-sort';
-import logo from '../assets/logo.svg';
 import runtimeEnv from '@mars/heroku-js-runtime-env';
-
-import PropTypes from "prop-types";
 
 import BetaBanner from "./BetaBanner";
 import Splash from "./Splash";
@@ -21,6 +11,10 @@ import ProgramDropdown from "./ProgramDropdown";
 import WardToggle from "./WardToggle";
 import SortDropdown from "./SortDropdown";
 import FeatureDetail from "./FeatureDetail";
+import FeatureList from "./FeatureList";
+import InteractiveMap from "./InteractiveMap";
+import MobileMapPopup from "./MobileMapPopup";
+
 import {
   BackToListViewButton,
   MobileListToggleButton,
@@ -29,445 +23,17 @@ import {
 } from "./Buttons";
 
 import * as constants from "../constants";
-import * as utils from "../utils";
 
-// Allows us to change envvars during runtime, without recompiling app on Heroku.
-// See: https://github.com/mars/create-react-app-buildpack/blob/master/README.md#runtime-configuration
+import logo from '../assets/img/logo.svg';
+
+import 'bootstrap/dist/css/bootstrap.min.css'; // Must come first.
+import 'simplebar/dist/simplebar.css';
+import "slick-carousel/slick/slick.css";
+import "slick-carousel/slick/slick-theme.css";
+import '../assets/scss/App.scss';
+
 const env = runtimeEnv()
 
-class FeatureListItem extends React.Component {
-  static propTypes = {
-    uid: PropTypes.number,
-    media: PropTypes.arrayOf(PropTypes.object),
-    artistName: PropTypes.string,
-    address: PropTypes.string,
-    year: PropTypes.number,
-    onClick: PropTypes.func,
-  }
-
-  static defaultProps = {
-    uid: 0,
-    media: [],
-  }
-
-  handleClick = () => {
-    ReactGA.event({
-      category: 'Artwork',
-      action: 'View details',
-      label: this.props.artistName,
-      value: this.props.uid,
-    })
-    this.props.onClick(this.props.uid)
-  }
-
-  render() {
-    const { media } = this.props
-    return (
-      <div className='lv-tile' onClick={this.handleClick}>
-        <div className='lv-tile-pic'>
-          <LazyLoad height={100} offset={30} overflow={true} resize={true}>
-            { /* eslint-disable-next-line jsx-a11y/img-redundant-alt */ }
-            <img
-              aria-label="Thumbnail Preview"
-              alt="Photo of artwork"
-              className="list-img"
-              src={utils.getCoverImage(media)}
-              onError={utils.handleMissingImage}
-            />
-          </LazyLoad>
-        </div>
-        <div className="lv-tile-txt">
-          <h5 className='tileArtist'>
-            {this.props.artistName}
-          </h5>
-          <p className='tileAddress'>
-            {this.props.address}
-          </p>
-          <p className='tileYear'>
-            {this.props.year}
-          </p>
-        </div>
-      </div>
-
-    );
-  }
-}
-class FeatureList extends React.Component {
-  static propTypes = {
-    features: PropTypes.arrayOf(PropTypes.object),
-    onItemClick: PropTypes.func,
-  }
-
-  componentDidUpdate() {
-    forceCheck()
-  }
-
-  render() {
-    return (
-      <div id="list">
-        {this.props.features.map(f =>
-          <FeatureListItem
-            key={f.uid}
-            uid={f.uid}
-            artistName={f.artist}
-            address={f.address}
-            year={f.yr}
-            media={f.media}
-            onClick={this.props.onItemClick}
-          />
-        )}
-      </div>
-    )
-  }
-}
-
-class GMap extends React.Component {
-  static propTypes = {
-    onFeatureMapClick: PropTypes.func,
-  }
-
-  state = {
-    prevActiveFeature: {},
-  }
-
-  getFeatureById = (featureId) => {
-    return this.map.data.getFeatureById(featureId)
-  }
-
-  render() {
-    return (
-      <div id="theMap">
-        <div id='map' ref="map"></div>
-      </div>
-    )
-  }
-
-  componentDidMount() {
-
-    // See: https://engineering.universe.com/building-a-google-map-in-react-b103b4ee97f1
-    const googleMapScript = document.createElement('script')
-    // We're using the default weekly channel, which is fine so long as we're
-    // using official library API calls. If there are ever more issues with
-    // random crashes when we didn't push any code changes ourselves, try
-    // locking the version to a previous numeric one.
-    // See: https://developers.google.com/maps/documentation/javascript/versions
-    googleMapScript.src = `https://maps.googleapis.com/maps/api/js?key=${env.REACT_APP_GOOGLE_MAPS_API_KEY}&v=weekly`
-    window.document.body.appendChild(googleMapScript)
-
-    googleMapScript.addEventListener('load', () => {
-      // create the map, marker and infoWindow after the component has
-      // been rendered because we need to manipulate the DOM for Google =(
-      this.map = this.createMap()
-      this.prepareMap();
-      this.map.data.loadGeoJson('geojson/ftrs.json', { idPropertyName: 'uid' })
-      this.map.data.loadGeoJson('geojson/wards.json', { idPropertyName: 'AREA_ID' })
-
-      this.map.data.addListener('click', (e)=> this.handleFeatureClick(e));
-    })
-  }
-
-  // clean up event listeners when component unmounts
-  componentDidUnMount() {
-    window.google.maps.event.clearListeners(this.map, 'zoom_changed')
-  }
-
-  createMap() {
-    let mapOptions = {
-      center: constants.DEFAULT_MAP_CENTER,
-      zoom: constants.MAP_ZOOM_LEVEL.DEFAULT,
-      mapTypeControl: false,
-      fullscreenControl: false,
-      mapTypeId: window.google.maps.MapTypeId.ROADMAP,
-      styles: constants.MAP_STYLE_BASE,
-    }
-    return new window.google.maps.Map(this.refs.map, mapOptions)
-  }
-
-  /**
-   * Iterates through Feature objects in FeatureCollecton, and checks properties
-   * for values matching the active options in the Select dropdown. If active
-   * for each filter (year, ward, program), makes feature visible and appends to
-   * "li" array; otherwise hides Feature.
-   *
-   * @param {Select.OptionsType} activeYearOpts -
-   * @param {Select.OptionsType} activeWardOpts -
-   * @param {Select.OptionsType} activeProgramOpts -
-   * @see https://react-select.com/props
-   *
-   * @returns {undefined}
-   */
-  filterMap(activeYearOpts, activeWardOpts, activeProgramOpts, setVisibleFeatures) {
-    let visibleFeatures = [];
-
-    let map = this.map;
-    map.data.forEach(function(feature) {
-      const checkForKeep = (feature, propName, activeOpts) => {
-        for (let i = 0; i < activeOpts.length; i++) {
-          if (feature.getProperty(propName) &&
-            feature.getProperty(propName).toString() === activeOpts[i].value.toString()
-          ) {
-            return true;
-          }
-        }
-        return false;
-      }
-
-      let keepForYear = checkForKeep(feature, 'yr', activeYearOpts)
-      let keepForWard = checkForKeep(feature, 'ward', activeWardOpts)
-      let keepForProgram = checkForKeep(feature, 'prgrm', activeProgramOpts)
-
-      const isArtwork = (feature) => (
-        feature.getGeometry() !== null &&
-        feature.getGeometry().getType() === 'Point'
-      )
-
-      if (isArtwork(feature)) {
-        if (keepForYear && keepForWard && keepForProgram) {
-          map.data.overrideStyle(feature, { visible: true });
-          visibleFeatures.push({
-            key: feature.getProperty('uid').toString(),
-            uid: feature.getProperty('uid'),
-            artist: feature.getProperty('artist'),
-            yr: feature.getProperty('yr'),
-            address: feature.getProperty('address'),
-            media: feature.getProperty("media"),
-          })
-        } else {
-          map.data.overrideStyle(feature, { visible: false });
-        }
-      }
-    })
-
-    setVisibleFeatures(visibleFeatures);
-  }
-
-  /**
-   * This handler deals with both clicks to map features (both ward and
-   * artwork), but also clicks to a FeatureListItem.
-   */
-  handleFeatureClick(e){
-    // If we're clicking on the map, the feature is in a property, otherwise
-    // it's the whole event object. TODO split into separate functions.
-    const isMapClickEvent = (!!e.feature)
-    let feature = isMapClickEvent ? e.feature : e
-
-    // Revert the previously selected feature's style.
-    this.map.data.revertStyle(this.state.prevActiveFeature);
-    // Required because default ward style is invisible.
-    this.map.data.overrideStyle(this.state.prevActiveFeature, {visible: true})
-
-    // Clicking on an artwork point feature on the map.
-    if (feature.getGeometry().getType() === "Point") {
-      const prgrm = feature.getProperty('prgrm');
-      if (prgrm !== "Partnership Program" && prgrm !==  "Outside the Box" && prgrm !==  "StART Support"){
-        feature.setProperty('prgrm', "Other");
-      };
-      this.map.data.overrideStyle(feature, {
-        // Ensure active marker always on top.
-        zIndex: 10000,
-        icon: constants.ICONS_LRG[prgrm].icon,
-      });
-    }
-
-    // Clicking a ward feature on the map.
-    if (feature.getGeometry().getType() === "MultiPolygon") {
-      // Clicking on a ward mulitpolygon feature on the map.
-      this.map.data.overrideStyle(feature, constants.MAP_STYLE_WARD_ACTIVE);
-    }
-
-    this.setState({ prevActiveFeature: feature })
-
-    if (isMapClickEvent) {
-      // Clicking a feature object on the map.
-      this.props.onFeatureMapClick(feature)
-    } else {
-      // Otherwise, must be FeatureListItem click.
-      // Pan to the LatLng object coordinates.
-      // See: https://stackoverflow.com/a/30130908
-      this.map.panTo(feature.getGeometry().get())
-      this.map.setZoom(constants.MAP_ZOOM_LEVEL.FEATURE);
-    }
-  };
-
-  geolocation(){
-    const map = this.map;
-    if (navigator.geolocation) {
-      navigator.geolocation.getCurrentPosition(function(position) {
-        let pos = {
-          lat: position.coords.latitude,
-          lng: position.coords.longitude
-        };
-        map.setCenter(pos);
-        map.setZoom(constants.MAP_ZOOM_LEVEL.FEATURE);
-      }, function() {
-        handleLocationError(true, map.getCenter());
-      });
-    } else {
-      // Browser doesn't support Geolocation
-      handleLocationError(false, map.getCenter());
-    }
-
-    function handleLocationError(browserHasGeolocation, pos) {
-
-    }
-  }
-
-  /**
-   * Set up the custom styling for our map.
-   */
-  prepareMap = () => {
-    this.map.data.setStyle(function(feature){
-      var geo = feature.getGeometry();
-      var prgrm = feature.getProperty('prgrm');
-      if (prgrm !== "Partnership Program" && prgrm !==  "Outside the Box" && prgrm !==  "StART Support"){
-        feature.setProperty('prgrm', "Other");
-      };
-      var type = "";
-      if (geo) {
-        type = geo.getType();
-      }
-
-      if (type === "MultiPolygon") {
-        return constants.MAP_STYLE_WARD_DEFAULT;
-      } else {
-        return ({
-          icon: constants.ICONS_REG[feature.getProperty('prgrm')].icon,
-          visible: true
-        });
-      }
-
-    })
-  }
-
-  prepareMapMobile() {
-    this.prepareMap()
-    this.map.setOptions({zoomControl:false, streetViewControl:false})
-  }
-
-  prepareMapDesktop() {
-    this.prepareMap()
-    this.map.setOptions({zoomControl:true, streetViewControl:true})
-  }
-
-  resetMap() {
-    this.map.panTo(constants.DEFAULT_MAP_CENTER);
-    this.map.setZoom(constants.MAP_ZOOM_LEVEL.DEFAULT);
-  }
-
-  showWardLayer = (show) => {
-    const isWard = (feature) => (
-      feature.getGeometry() &&
-      feature.getGeometry().getType() === "MultiPolygon"
-    )
-
-    this.map.data.forEach( (feature) => {
-      if (isWard(feature)) {
-        this.map.data.overrideStyle(feature, { visible: show })
-      }
-    })
-  }
-}
-
-/**
- * Renders the little hovering popup at the bottom of mobile map view, in
- * which ward and artwork data are shown when feature is clicked/active.
- */
-const MobileMapPopup = ({ onClick, activeFeature }) => {
-  if (typeof activeFeature.getProperty === 'undefined' ) { return null }
-
-  const getFeatureCoverImageSrc = () => {
-    if (activeFeature.getProperty('media')) {
-      return utils.getCoverImage(activeFeature.getProperty('media'))
-    }
-    return ''
-  }
-
-  // Only wards have this property.
-  const isArtwork = (!activeFeature.getProperty('AREA_L_CD'))
-
-  return (
-    <div id="MobileMapPopUp" onClick={onClick}>
-      { isArtwork ? (
-        <MobileMapPopupArtwork
-          imgSrc={getFeatureCoverImageSrc()}
-          year={activeFeature.getProperty('yr')}
-          artist={activeFeature.getProperty('artist')}
-          address={activeFeature.getProperty('address')}
-        />
-      ) : (
-        <MobileMapPopupWard
-          wardNumber={activeFeature.getProperty('AREA_L_CD')}
-          wardName={activeFeature.getProperty('AREA_NAME')}
-        />
-      )}
-    </div>
-  )
-}
-MobileMapPopup.propTypes = {
-  onClick: PropTypes.func.isRequired,
-  activeFeature: PropTypes.object,
-}
-MobileMapPopup.defaultProps = {
-  activeFeature: {},
-}
-
-/**
- * Component for rendering map popup when active feature geometry is a POLYGON
- * representing a WARD.
- */
-const MobileMapPopupWard = ({ wardNumber, wardName }) => {
-  return (
-    <div className="popup-txt">
-      <h5 className='detailWard'>
-        Ward {wardNumber} <br/>
-        {wardName}
-      </h5>
-    </div>
-  )
-}
-MobileMapPopupWard.propTypes = {
-  wardNumber: PropTypes.number.isRequired,
-  wardName: PropTypes.string.isRequired,
-}
-
-/**
- * Component for rendering map popup when active feature geometry is a POINT
- * representing an ARTWORK.
- */
-const MobileMapPopupArtwork = ({ imgSrc, year, artist, address }) => {
-  return (
-    <React.Fragment>
-      <div className='popup-pic'>
-        { /* eslint-disable-next-line jsx-a11y/img-redundant-alt */ }
-        <img alt="Photo of artwork" aria-label="Thumbnail Preview" src={imgSrc} className="list-img" onError={utils.handleMissingImage}/>
-      </div>
-      <div className="popup-txt">
-        <p>
-          <strong className='tileArtist'>
-            {artist}
-          </strong>
-        </p>
-        <p className='tileAddress'>
-          {address}
-        </p>
-        <p className='tileYear'>
-          Created in {year}
-        </p>
-      </div>
-    </React.Fragment>
-  )
-}
-MobileMapPopupArtwork.propTypes = {
-  imgSrc: PropTypes.string,
-  artist: PropTypes.string,
-  address: PropTypes.string,
-  year: PropTypes.number.isRequired,
-}
-MobileMapPopupArtwork.defaultProps = {
-  imgSrc: '',
-  artist: '',
-  address: '',
-}
 
 export default class App extends React.Component {
 
@@ -808,7 +374,7 @@ export default class App extends React.Component {
       <div className="parent">
         { showSplash ? <Splash onButtonClick={this.closeSplash} isMobile={isMobileView} /> : null }
         <BetaBanner isMobile={isMobileView}/>
-        <GMap onFeatureMapClick={this.handleMapClick} ref="mapControl" />
+        <InteractiveMap onFeatureMapClick={this.handleMapClick} ref="mapControl" />
         <GeolocateButton onClick={this.handleGeolocate}/>
 
         <div id="nav">
