+++ src/components/WardDropdown.js
@@ -1,38 +1,37 @@
-import React, { Component } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import SelectAll from "./SelectAll";
 import ReactGA from 'react-ga';
 
-import * as constants from "../constants";
+import { WARD_OPTS } from "../constants";
 
-class WardDropdown extends Component {
-  static propTypes = {
-    onSelect: PropTypes.func,
-    selected: PropTypes.array,
-  };
+const WardDropdown = ({ onSelect, selected }) => {
 
-  handleChange = (selectedOptions) => {
+  const handleChange = (selectedOptions) => {
     ReactGA.event({
       category: 'Form Fields',
       action: 'Filter artwork',
       label: 'By ward',
     })
-    this.props.onSelect(selectedOptions);
+    onSelect(selectedOptions);
   };
 
-  render() {
-    return (
-      <SelectAll
-        allowSelectAll = {true}
-        closeMenuOnSelect={false}
-        isMulti = {true}
-        value={this.props.selected}
-        onChange = {this.handleChange}
-        options = {constants.WARD_OPTS}
-        className={"drp"}
-      />
-    );
-  }
+  return (
+    <SelectAll
+      allowSelectAll = {true}
+      closeMenuOnSelect={false}
+      isMulti = {true}
+      value={selected}
+      onChange = {handleChange}
+      options = {WARD_OPTS}
+      className={"drp"}
+    />
+  );
+};
+
+WardDropdown.propTypes = {
+  onSelect: PropTypes.func,
+  selected: PropTypes.array,
 };
 
 export default WardDropdown;
